Перераспределил структуры и функции?
Структуры Head и Deputy (наследники от Employee по заданию) так должны выглядеть?
